requirements.txt
pypresence==3.2.1
python-dotenv~=0.21.0